requirement.txt
backcall==0.1.0
beautifulsoup4==4.5.3
chardet==3.0.4
colorama==0.3.9
comtypes==1.1.7
decorator==4.3.0
et-xmlfile==1.0.1
jdcal==1.4
jedi==0.12.0
nose==1.3.7
numpy==1.15.1
openpyxl==2.5.5
pandas==0.23.4
parameterized==0.6.1
parso==0.2.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
pywin32==223
pywinauto==0.6.5
selenium==3.13.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
uiautomation==1.1.8
wcwidth==0.1.7
xlrd==1.1.0
